requirements.txt
Django==5.0
psycopg2==2.9.5
gunicorn==20.1.0
django-storages==1.13.1
boto3==1.26.33
pillow==9.4.0
qrcode==7.4.2
asgiref==3.7.2
django-cors-headers==4.3.1
sqlparse==0.4.4
djangorestframework==3.14.0